pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rehablarte"
version = "0.1.0"
description = "Telegram bot that will reply sent message as a voice note"
requires-python = ">=3.12"
dependencies = [
    "aiogram==3.23.0",
    "python-dotenv",
    "piper-tts",     # Lightweight TTS (~50MB)
    "onnxruntime",
    "faster-whisper",
    "loguru"
]

[project.scripts]
rehablarte = "rehablarte.main:main"

[tool.setuptools.packages.find]
where = ["src"]